+++ Changelog.md
@@ -11,6 +11,7 @@
 - RSML without an image can now be correctly displayed on the viewer
 - Added new 'absolute average' measurements for angles
 - Checking a node in the add button folder picker now checks and expands all children
+- Selecting a model in the add button folder picker now propagates that model to all checked children
 
 # 0.6.0:
 - Added auto-updating. App will download, notify, and update on close